JG 20250505 Refactor controllers and mapping logic; add helper utilities and update response models

+++ ToDoAPI/Controllers/TasksController.cs
@@ -5,8 +5,10 @@ using Serilog;
 using System.Security.Claims;
 using System.Threading.Tasks;
 using ToDoAPI.Entities;
+using ToDoAPI.Helpers;
 using ToDoAPI.Interfaces.Repository;
 using ToDoAPI.Models.Mapper;
+using ToDoAPI.Models.Reponse;
 using ToDoAPI.Models.Request;
 using ToDoAPI.Sevices;
 
@@ -14,35 +16,32 @@ namespace ToDoAPI.Controllers
 {
     [Route("api/[controller]")]
     [ApiController]
-    public class TasksController : ControllerBase
+    public class TasksController : AuthenticatedControllerBase
     {
         private readonly ITaskRepository _repository;
-        private readonly IUserRepository _repositoryUser;
-     
+    
 
-        public TasksController(ITaskRepository repository, IUserRepository repositoryUser )
+        public TasksController(ITaskRepository repository)
         {
             _repository = repository;
-            _repositoryUser = repositoryUser;
+            
         }
 
 
         [Authorize]
         [HttpPost]
-        public async Task<ActionResult> Create(TaskItemRequest request)
+        public async Task<ActionResult> Create([FromBody] TaskItemRequest request)
         {
             try
             {
-
-                var userId = User.FindFirst(ClaimTypes.NameIdentifier);
-
                 var newTask = request.ToTaskItem();
 
-                newTask.UserId = Guid.Parse(userId!.Value);
+                newTask.UserId = UserId;
 
-                await _repository.AddAsync(newTask);
+                var response = await _repository.AddAsync(newTask);
 
-                return CreatedAtAction(nameof(Create),new {request.Title},request);
+                return CreatedAtAction(nameof(GetById), new { id = newTask.Id }, response);
+
             }
             catch (Exception ex)
             {
@@ -55,43 +54,65 @@ namespace ToDoAPI.Controllers
 
         [Authorize]
         [HttpGet]
-        public async Task<ActionResult> GetAll()
+        public async Task<ActionResult<IEnumerable<TaskItemReponse>>> GetAll()
         {
             try
             {
 
-                var userId = User.FindFirst(ClaimTypes.NameIdentifier);
+                var tasks = await _repository.GetAllByUser(UserId);
 
-                var Id = Guid.Parse(userId!.Value);
+                var taskMapped = tasks.ToResponse();
 
-                var tasks = await _repository.GetAllByUser(Id);
-
-                return Ok(tasks);
+                return Ok(taskMapped);
             }
             catch (Exception ex)
             {
 
                 var userId = User.FindFirst(ClaimTypes.NameIdentifier)?.Value ?? "Desconocido";
-                Log.Error(ex, "Ocurrió un error al cargar la lista de tareas del empleado {Id}", userId);
+                Log.Error(ex, "Ocurrió un error al cargar la lista de tareas del empleado {Id}", UserId);
 
                 return Problem();
             }
         }
 
-
         [Authorize]
-        [HttpDelete]
-        public async Task<ActionResult> Delete([FromQuery]Guid id)
+        [HttpGet("{id}")]
+        public async Task<ActionResult<TaskItemReponse>> GetById([FromRoute] Guid id)
         {
             try
             {
 
                 var task = await _repository.GetByIdAsync(id);
 
-                if (task is null)
-                {
+
+                if (task is null || task.UserId != UserId)
                     return NotFound("No se encontro la tarea ");
-                }
+                
+
+
+                var taskMapped = task.ToResponse();
+
+                return Ok(taskMapped);
+            }
+            catch (Exception ex)
+            {
+                Log.Error(ex, "Ocurrió un error al cargar la lista de tareas del empleado {Id}", UserId);
+
+                return Problem();
+            }
+        }
+
+        [Authorize]
+        [HttpDelete("{id}")]
+        public async Task<ActionResult> Delete([FromRoute]Guid id)
+        {
+            try
+            {
+
+                var task = await _repository.GetByIdAsync(id);
+
+                if (task is null || task.UserId != UserId)
+                    return NotFound("No se encontró la tarea o no tiene permiso para eliminarla");
 
 
                 await _repository.DeleteAsync(task);
@@ -108,17 +129,17 @@ namespace ToDoAPI.Controllers
         }
 
         [Authorize]
-        [HttpPut]
-        public async Task<ActionResult> Update([FromQuery]Guid id, [FromBody] TaskItemRequest request)
+        [HttpPut("{id}")]
+        public async Task<ActionResult> Update([FromRoute]Guid id, [FromBody] TaskItemRequest request)
         {
             try
             {
                 var task = await _repository.GetByIdAsync(id);
 
-                if (task is null)
-                {
-                    return NotFound("No se encontro la tarea");
-                }
+
+                if (task is null || task.UserId != UserId)
+                    return NotFound("No se encontró la tarea o no tiene permiso para actualizarla");
+
 
                 request.MapTo(task);
